backend/requirements.txt
googlemaps~=4.10.0
pandas~=2.2.3
Flask~=3.0.3
protobuf~=5.26.1
Werkzeug~=3.0.6
google-maps-routeoptimization~=0.1.7
openpyxl==3.1.2
reportlab==4.1.0
pypdf==3.17.0
python-dotenv==1.0.1
gunicorn==23.0.0